Removed dead code based on benchmark persistence
